Improve test coverage of vuln management upload tasks
